requirements.txt
Django>=5.0
djangorestframework>=3.12,<4.0
celery>=5.2,<6.0
djoser>=2.1,<3.0